requirements.txt
Flask==0.10.1
Flask-Admin==1.1.0
Flask-Assets==0.10
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==1.3.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-User==0.6.1
Flask-WTF==0.11
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
Pillow==2.8.1
SQLAlchemy==0.9.8
WTForms==2.0.2
Werkzeug==0.10
alembic==0.7.4
argparse==1.2.1
blinker==1.3
closure==20140110
gunicorn==19.3.0
html-tree-diff==0.1.2
html5lib==0.999
itsdangerous==0.24
lxml==3.4.2
passlib==1.6.2
psycopg2==2.6
py-bcrypt==0.4
pycrypto==2.6.1
pyenchant==1.6.6
reportlab==3.1.44
requests==2.5.3
six==1.9.0
validate-email==1.2
webassets==0.10.1
wsgiref==0.1.2
